% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_i.R
\name{fa_i}
\alias{fa_i}
\title{Generate a Font Awesome \verb{<i>} tag}
\usage{
fa_i(name, class = NULL, htmlDependency = NULL, ...)
}
\arguments{
\item{name}{The name of the Font Awesome icon.}

\item{class}{Additional classes to customize the style of the icon (see the
usage examples for details on supported styles).}

\item{htmlDependency}{Provides an opportunity to use a custom
\code{html_dependency} object (created via a call to
\code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}) instead of one supplied by the function
(which uses Font Awesome's free assets and are bundled in the package). A
custom \code{html_dependency} object is useful when you have paid icons from
Font Awesome or would otherwise like to customize exactly which icon assets
are used (e.g., woff, woff2, eot, etc.). By default, this is \code{NULL} where
the function interally generates an \code{html_dependency}.}

\item{...}{Arguments passed to the \verb{<i>} tag of \link[htmltools:builder]{htmltools::tags}.}
}
\value{
An icon element.
}
\description{
The \code{fa_i()} function creates a Font Awesome \verb{<i>} tag and not an SVG as with
\code{\link[=fa]{fa()}}. The primary use case for \code{fa_i()} is for legacy Shiny applications
that use the \code{\link[shiny:icon]{shiny::icon()}} function. This function is called within a
\code{\link[shiny:icon]{shiny::icon()}} call and all HTML dependencies to support icon generation are
hosted in the \strong{fontawesome} package.
}
\examples{
if (interactive()) {

# Create a Font Awesome icon object
fa_i(name = "r-project")

}

}
