% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_info.R
\name{fa_info}
\alias{fa_info}
\title{Generate a Font Awesome \verb{<i>} tag}
\usage{
fa_info()
}
\value{
Nothing is returned.
}
\description{
The \code{fa_info()} function will display an interactive table that contains all
Font Awesome icons included in the package. The first column displays the
icon itself and the second provides the classification of icon style where
\code{"S"} is solid, \code{"R"} is regular, and \code{"B"} signifies that the icon is for a
brand.

The \code{"Icon Name"} column provides the short name and \code{"Full Name"} is the
fully-qualified, unique name for the icon (incorporating the type of icon,
since icons may have multiple styles). Both the short name and the full name
are acceptable inputs to \code{\link[=fa]{fa()}}, \code{\link[=fa_i]{fa_i()}}, and \code{\link[=fa_png]{fa_png()}}. In the case where
an icon has multiple styles and a short name is supplied to any of the
aforementioned functions, the solid (\code{"S"}) style will be used.
}
\examples{
if (interactive()) {

# View a table of all Font Awesome icons in the package
fa_info()

}

}
