% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_metadata.R
\name{fa_metadata}
\alias{fa_metadata}
\title{Get metadata on the included Font Awesome assets}
\usage{
fa_metadata()
}
\value{
A list with metadata for the included Font Awesome assets.
}
\description{
This function provide some metadata about the included Font Awesome assets
in the \strong{fontawesome} package. The list that is returned has the following
components:
\itemize{
\item \code{version}: The released version number for the Font Awesome icons
\item \code{icon_count}: The total count of unique Font Awesome icons
\item \code{icon_names}: A vector of short names (e.g., \code{"npm"}, \code{"drum"}, etc.) for
all included icons
\item \code{icon_names_full}: A vector containing the full names (e.g., \code{"fab fa-npm"}, \code{"fas fa-drum"}, etc.) for all included icons
\item \code{icon_names_fa(r|s|b)}: Vectors of short names within the regular (\code{"r"}),
solid (\code{"s"}), and brand (\code{"b"}) groups
\item \code{icon_names_full_fa(r|s|b)}: Vectors with the full names of icons within
the regular (\code{"r"}), solid (\code{"s"}), and brand (\code{"b"}) groups
}
}
\examples{
if (interactive()) {

# Get information on the Font Awesome
# assets included in this package
fa_metadata()

}

}
