% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_svg.R
\name{fa_svg}
\alias{fa_svg}
\title{Generate a FontAwesome icon as an SVG}
\usage{
fa_svg(name, height = 30, fill = NULL)
}
\arguments{
\item{name}{the name of the FontAwesome icon.}

\item{height}{the absolute height (px)
of the rendered SVG.}

\item{fill}{an option to change the fill
color of the icon.}
}
\value{
a character object that is the styled
FontAwesome icon within \code{svg} tags.
#' @examples
# Emit a FontAwesome icon (`r-project`) as
# SVG within `svg` tags
fa_svg(name = "r-project")
}
\description{
Add a FontAwesome icon as SVG contained within
\code{<svg>...</svg>}. We can optionally set
certain style attributes.
}
