% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fai.R
\name{fai}
\alias{fai}
\title{Generate an inline FontAwesome icon}
\usage{
fai(name)
}
\arguments{
\item{name}{the name of the FontAwesome icon.}

\item{fill}{an option to change the fill
color of the icon.}
}
\value{
a character object that is the styled
FontAwesome icon within \code{svg} tags.
}
\description{
Add a FontAwesome icon within inline text.
We can optionally set certain style attributes.
}
\examples{
# Emit a FontAwesome icon (`r-project`) as
# SVG within `i` tags
fa(name = "r-project")
}
