% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fontawesome}
\alias{print.fontawesome}
\title{Print the fontawesome icon to the Viewer}
\usage{
\method{print}{fontawesome}(x, view = interactive(), ...)
}
\arguments{
\item{x}{An agent object of class \code{fontawesome}.}

\item{view}{The value for \code{print()}s \code{browse} argument.}

\item{...}{Any additional parameters.}
}
\value{
No return value, called for printing to the Viewer.
}
\description{
This function will show the fontawesome icon in the Viewer.
}
\keyword{internal}
