% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa.R
\name{sfa}
\alias{sfa}
\title{Generate a FontAwesome icon as an SVG for Shiny}
\usage{
sfa(name, height = NULL, fill = NULL)
}
\arguments{
\item{name}{the name of the FontAwesome icon.}

\item{height}{the absolute height (px)
of the rendered SVG.}

\item{fill}{an option to change the fill
color of the icon.}
}
\value{
a character object that is the styled
FontAwesome icon within \code{svg} tags.
}
\description{
Add a FontAwesome icon to a Shiny app as SVG
contained within \code{<svg>...</svg>}. We
can optionally set certain style attributes.
This function differs from \code{fa()} in that
it is specialized for Shiny apps by preserving
the output as HTML.
}
\examples{
# Emit a FontAwesome icon (`r-project`) as
# SVG within `svg` tags
sfa(name = "r-project")
}
