% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_i.R
\name{fa_i}
\alias{fa_i}
\title{Generate a FontAwesome \verb{<i>} tag}
\usage{
fa_i(name, class = NULL, ...)
}
\arguments{
\item{name}{The name of the FontAwesome icon.}

\item{class}{Additional classes to customize the style of the
icon (see the usage examples for details on supported styles).}
}
\description{
Generate a FontAwesome \verb{<i>} tag
}
