% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.R
\name{knit_print.fontawesome}
\alias{knit_print.fontawesome}
\title{Knit the FontAwesome icon}
\usage{
\method{knit_print}{fontawesome}(x, ...)
}
\arguments{
\item{x}{an object containing the class
\code{fontawesome}.}
}
\description{
This facilitates printing of
the FontAwesome icon within R Markdown.
}
\keyword{internal}
